feat: move search logic to custom search hook
