+++ src/components/Dashboard/Dashboard.tsx
@@ -1,24 +1,15 @@
-import { FC, useCallback, useRef, useState } from 'react';
+import { FC } from 'react';
 
 import Styles from './Dashboard.module.css';
 import { search } from '../../helpers/search';
+import { useCustomSearch } from '../../hooks/useCustomSearch';
 import { DataProp } from '../../types/DataProp';
-import { logging } from '../../utils/logging';
 import { Loader } from '../Loader/Loader';
 import { Results } from '../Results/Results';
 import { Search } from '../Search/Search';
 
 export const Dashboard: FC<DataProp> = ({ datasites, datatests }): JSX.Element => {
-  const inputRef = useRef<HTMLInputElement>(null);
-  const [getValue, setValue] = useState<string>('');
-  const callvalue = useCallback((value: string): void => {
-    setValue(value);
-    logging(value);
-  }, []);
-  const resetvalue = useCallback((): void => {
-    inputRef.current!.value = '';
-    setValue('');
-  }, []);
+  const { inputRef, getValue, callvalue, resetvalue } = useCustomSearch();
   return (
     <>
       <h1 className={Styles.title}>Dashboard</h1>
